extract translation string syncing to it’s own module
