+++ app/controllers/api/v1/projects_controller.rb
@@ -2,6 +2,7 @@ class Api::V1::ProjectsController < Api::ApiController
   include FilterByOwner
   include FilterByCurrentUserRoles
   include TranslatableResource
+  include SyncResourceTranslationStrings
   include IndexSearch
   include FilterByTags
   include AdminAllowed
